Renamed Editions "article" to "layout"
- Consistent with dotcom and apps terminology
- Renamed stories to `Editions/Layouts`
